% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlUtils.R
\name{dmlTab}
\alias{dmlTab}
\title{Table maker for DML}
\usage{
dmlTab(..., modsel = F)
}
\arguments{
\item{...}{list of \code{plrFit} or \code{irmFit} output lists of arbitrary length}

\item{modsel}{Boolean for whether to return indices of best outcome and treatment models}
}
\description{
Table maker for DML
}
