% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmlUtils.R
\name{plrFit}
\alias{plrFit}
\title{Wrapper for Partially Linear Model}
\usage{
plrFit(dat, outcomeMod, pscoreMod, nf = 3)
}
\arguments{
\item{dat}{A DoubleMLData object initialised using using \code{DoubleMLData$new}}

\item{outcomeMod}{A mlr3 learner object for the outcome model}

\item{pscoreMod}{A mlr3 learner object for the treatment model}

\item{nf}{Number of folds for cross-fitting}
}
\value{
list with (estimate, SE, outcome RMSE, treatment RMSE, treatment Classification Error)
}
\description{
Wrapper for Partially Linear Model
}
\references{
Bach, P., V. Chernozhukov, M. S. Kurz, and M. Spindler. (2021):
“DoubleML -- An Object-Oriented Implementation of Double Machine Learning
in R,” arXiv \link{stat.ML},.
}
